requirements.txt
Django==4.2.7
django-crispy-forms==2.0
crispy-bootstrap5==0.7
django-filter==23.3
django-import-export==3.3.1
openpyxl==3.1.2
pandas==2.1.3
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0